refactor: 이름 변경 및 코드 리팩토링
